Create The Adventures of ACHIMA
It is about Amelia, Chloe, Hannah, Ivanna, Maddison and Avery. They are from different states in America, and on the first day of school, they accidentally travel to the fairy realm and they are stuck there. How will they get out? Who and what will they meet? This is the Adventure of the ACHIMA!

Enjoy :)